requirements.txt
requests
google-api-python-client
google-cloud-recommender